pyproject.toml
[project]
name = "rag-multidoc-agent"
version = "0.1.0"
description = "Un agente RAG para procesar y consultar múltiples documentos PDF"
authors = [
    {name = "raistln",email = "samumarfon@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "dotenv (>=0.9.9,<0.10.0)",
    "langchain (>=0.3.23,<0.4.0)",
    "llama-index-core (>=0.12.30,<0.13.0)",
    "langchain-huggingface (>=0.1.2,<0.2.0)",
    "langchain-mistralai (>=0.2.10,<0.3.0)",
    "llama-index-llms-langchain (>=0.6.1,<0.7.0)",
    "llama-index (>=0.12.30,<0.13.0)",
    "llama-index-embeddings-huggingface (>=0.5.3,<0.6.0)",
    "langchain-community (>=0.3.21,<0.4.0)",
    "llama-index-llms-mistralai (>=0.4.0,<0.5.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
]

[tool.poetry]
name = "rag-multidoc-agent"
version = "0.1.0"
description = "Un agente RAG para procesar y consultar múltiples documentos PDF"
authors = ["Tu Nombre <tu.email@ejemplo.com>"]
readme = "README.md"
packages = [{include = "rag_multidoc_agent", from = "src"}]


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"
pytest = "^7.4.0"
black = "^23.7.0"
isort = "^5.12.0"
mypy = "^1.5.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[project.urls]
"Homepage" = "https://github.com/tu-usuario/rag-multidoc-agent"
"Bug Tracker" = "https://github.com/tu-usuario/rag-multidoc-agent/issues"
